reconstrução dos menus e implementação do file manager

+++ Treinamento Interfaces - Exercicio 2/src/orchestrator/FileOrchestrator.java
@@ -22,7 +22,6 @@ public class FileOrchestrator extends FolderOrchestrator implements ImageFileDat
         try {
             URL url = new URL(content);
             image = ImageIO.read(url);
-            System.out.println(directory);
             ImageIO.write(image, "jpeg", new File(directory + nameFile));
         } catch (IOException e) {
             e.printStackTrace();
@@ -34,12 +33,12 @@ public class FileOrchestrator extends FolderOrchestrator implements ImageFileDat
 
     public void saveAllListOfImagesFiles(List<MFile> mImageFileList) {
         for (MFile lista : mImageFileList) {
-            Scanner tc = new Scanner(System.in);
             System.out.println("DIRECTORY PATH");
             String diretorio = tc.nextLine();
-            diretorio +="/image/";
+            diretorio += "/image/";
 
             file = new File(diretorio);
+            System.out.println(diretorio);
 
             if (file.exists()) {
 
@@ -49,9 +48,54 @@ public class FileOrchestrator extends FolderOrchestrator implements ImageFileDat
         }
     }
 
+    public void saveAllListOfFiles(List<MFile> mFileList) {
+        for (MFile lista : mFileList) {
+            System.out.println("DIRECTORY PATH");
+            String diretorio = tc.nextLine();
+            file = new File(diretorio);
+            if (file.exists()) {
+                saveFile(diretorio, lista.getContent(), lista.getType(), lista.getNameFile());
+            } else System.out.println("DIRECTORY NOT FOUND");
+
+        }
+
+    }
+
+    @Override
+    public void saveFile(String directory, String content, MFileAnnotationTypeEnum type, String nameFile) {
+        String subdirectory;
+        if (type.equals(MFileAnnotationTypeEnum.REMINDER)){
+            System.out.println("entrei no reminder");
+            directory += "/reminder";
+        }
+        else if (type.equals(MFileAnnotationTypeEnum.IMPORTANT)) {
+            directory +="/important";
+
+        } else if(type.equals((MFileAnnotationTypeEnum.SIMPLE))){
+            directory = directory;
+        } else{
+            System.out.println("TYPE NOT EXIST");
+        }
+
+        file = new File(directory);
+        if (file.exists()) {
+            directory = file + "/"+nameFile;
+
+            try {
+                BufferedWriter escreverTxt = new BufferedWriter(new FileWriter(directory));
+                escreverTxt.write(content);
+                escreverTxt.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        } else System.out.println("DIRECTORY NOT FOUND");
+
+
+    }
+
     @Override
     public void listAllImageFiles(String directory) {
-        file = new File(directory);
+        file = new File(directory + "/image");
         File afile[] = file.listFiles();
         int i = 0;
         for (int j = afile.length; i < j; i++) {
@@ -65,11 +109,11 @@ public class FileOrchestrator extends FolderOrchestrator implements ImageFileDat
 
     @Override
     public void removeImageFile(String directory, String nameFile) {
-        file = new File(directory+nameFile);
+        file = new File(directory + nameFile);
 
-        if(file.isFile()){
+        if (file.isFile()) {
             file.delete();
-        }else System.out.println("NOT FOUND IMAGE");
+        } else System.out.println("NOT FOUND IMAGE");
 
 //        File [] aFile = file.listFiles();
 //        for (File listFile : aFile) {
@@ -89,20 +133,6 @@ public class FileOrchestrator extends FolderOrchestrator implements ImageFileDat
     }
 
     @Override
-    public void saveFile(String directory, String content, MFileAnnotationTypeEnum type, String nameFile)  {
-        file = new File(directory + nameFile);
-        try {
-            BufferedWriter escreverTxt = new BufferedWriter(new FileWriter (file));
-            escreverTxt.write(content);
-            escreverTxt.close();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-
-    }
-
-    @Override
     public void recoveryFile(String directory, String nameFile) {
 
 
@@ -111,7 +141,10 @@ public class FileOrchestrator extends FolderOrchestrator implements ImageFileDat
     @Override
     public void removeFile(String directory, String nameFile) {
         file = new File(directory + nameFile);
-        file.delete();
+        if(file.exists()) {
+            file.delete();
+        } else System.out.println("NOT FOUND");
+
 
     }
